--- src/dymoprint/__init__.py
@@ -1,4 +1,0 @@
-from dymoprint.lib.dymo_labeler import DymoLabeler
-from dymoprint.metadata import __version__
-
-__all__ = ["__version__", "DymoLabeler"]
