+++ src/labelle/__init__.py
@@ -1,0 +1,4 @@
+from labelle.lib.dymo_labeler import DymoLabeler
+from labelle.metadata import __version__
+
+__all__ = ["__version__", "DymoLabeler"]
