+++ src/labelle/_vendor/matplotlib.pyi
