+++ src/labelle/_vendor/matplotlib/LICENSE_BAKOMA
