+++ src/labelle/_vendor/matplotlib/LICENSE_JSXTOOLS_RESIZE_OBSERVER
