+++ src/labelle/_vendor/matplotlib/LICENSE_QT4_EDITOR
