+++ src/labelle/_vendor/matplotlib/LICENSE_STIX
