+++ src/labelle/lib/barcode_writer.py
