+++ src/labelle/lib/detect.py
