+++ src/labelle/lib/render_engines/barcode_with_text.py
@@ -5,13 +5,13 @@ from typing import Literal
 
 from PIL import Image
 
-from dymoprint.lib.render_engines.barcode import BarcodeRenderEngine
-from dymoprint.lib.render_engines.render_context import RenderContext
-from dymoprint.lib.render_engines.render_engine import (
+from labelle.lib.render_engines.barcode import BarcodeRenderEngine
+from labelle.lib.render_engines.render_context import RenderContext
+from labelle.lib.render_engines.render_engine import (
     RenderEngine,
     RenderEngineException,
 )
-from dymoprint.lib.render_engines.text import TextRenderEngine
+from labelle.lib.render_engines.text import TextRenderEngine
 
 
 class BarcodeWithTextRenderEngine(RenderEngine):
