+++ src/labelle/lib/render_engines/empty.py
@@ -1,7 +1,7 @@
 from PIL import Image
 
-from dymoprint.lib.render_engines.render_context import RenderContext
-from dymoprint.lib.render_engines.render_engine import RenderEngine
+from labelle.lib.render_engines.render_context import RenderContext
+from labelle.lib.render_engines.render_engine import RenderEngine
 
 
 class EmptyRenderEngine(RenderEngine):
