+++ src/labelle/lib/render_engines/horizontally_combined.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 from PIL import Image
 
-from dymoprint.lib.render_engines.empty import EmptyRenderEngine
-from dymoprint.lib.render_engines.render_context import RenderContext
-from dymoprint.lib.render_engines.render_engine import RenderEngine
+from labelle.lib.render_engines.empty import EmptyRenderEngine
+from labelle.lib.render_engines.render_context import RenderContext
+from labelle.lib.render_engines.render_engine import RenderEngine
 
 
 class HorizontallyCombinedRenderEngine(RenderEngine):
