+++ src/labelle/lib/render_engines/picture.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 from PIL import Image, ImageOps
 
-from dymoprint.lib.render_engines.render_context import RenderContext
-from dymoprint.lib.render_engines.render_engine import (
+from labelle.lib.render_engines.render_context import RenderContext
+from labelle.lib.render_engines.render_engine import (
     RenderEngine,
     RenderEngineException,
 )
