+++ src/labelle/lib/render_engines/print_payload.py
@@ -4,9 +4,9 @@ from typing import Literal
 
 from PIL import Image
 
-from dymoprint.lib.render_engines.margins import MarginsRenderEngine
-from dymoprint.lib.render_engines.render_context import RenderContext
-from dymoprint.lib.render_engines.render_engine import RenderEngine
+from labelle.lib.render_engines.margins import MarginsRenderEngine
+from labelle.lib.render_engines.render_context import RenderContext
+from labelle.lib.render_engines.render_engine import RenderEngine
 
 
 class PrintPayloadRenderEngine(RenderEngine):
