+++ src/labelle/lib/render_engines/print_preview.py
@@ -5,10 +5,10 @@ from typing import Literal
 from darkdetect import isDark
 from PIL import Image, ImageColor, ImageDraw, ImageOps
 
-from dymoprint.lib.render_engines.margins import MarginsRenderEngine
-from dymoprint.lib.render_engines.render_context import RenderContext
-from dymoprint.lib.render_engines.render_engine import RenderEngine
-from dymoprint.lib.utils import px_to_mm
+from labelle.lib.render_engines.margins import MarginsRenderEngine
+from labelle.lib.render_engines.render_context import RenderContext
+from labelle.lib.render_engines.render_engine import RenderEngine
+from labelle.lib.utils import px_to_mm
 
 
 class PrintPreviewRenderEngine(RenderEngine):
