+++ src/labelle/lib/render_engines/text.py
@@ -5,9 +5,9 @@ from typing import Literal
 
 from PIL import Image, ImageFont
 
-from dymoprint.lib.render_engines.render_context import RenderContext
-from dymoprint.lib.render_engines.render_engine import RenderEngine
-from dymoprint.lib.utils import draw_image
+from labelle.lib.render_engines.render_context import RenderContext
+from labelle.lib.render_engines.render_engine import RenderEngine
+from labelle.lib.utils import draw_image
 
 
 class TextRenderEngine(RenderEngine):
