+++ src/labelle/resources/icons/barcode_text_icon.png
