+++ src/labelle/resources/icons/logo_small.png
