+++ src/labelle/resources/icons/qr_icon.png
