+++ src/labelle/resources/icons/txt_icon.png
